server/requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
dnspython==1.16.0
eventlet==0.29.1
fake-useragent==0.1.11
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
parse==1.18.0
Pillow==8.0.1
pycodestyle==2.6.0
pyee==7.0.4
pylint==2.6.0
pyppeteer==0.2.2
pyquery==1.4.1
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
requests-html==0.10.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tqdm==4.51.0
typed-ast==1.4.1
urllib3==1.25.11
w3lib==1.22.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
